Invoke component lifecycle hooks predictably
Previously, lifecycle hooks on components (e.g. `didInsertElement`) were
invoked on components in an ad-hoc and unpredictable way.

In this commit, we introduce the notion of two types of lifecycle hooks:

1. Immediately-dispatched hooks for teardown, which are invoked before
   the state of the component changes so that the user can teardown any
   setup they had done previously.
2. After-render hooks, which are dispatched bottom-up and
   “asynchronously” (i.e., after the entire render process has
   completed) and are intended for new setup after the state of the
   component has changed.

This is possible because there are only two entry points that can
result in DOM mutations, both of which live in the Renderer. This means
that we can populate a list of lifecycle hooks to invoke as we walk the
render tree, and flush it when we are done.

No matter how many discrete changes *cause* DOM mutations, only a single
call to `renderer.revalidateTopLevelView` can be made at a time, and
once revalidation has been scheduled, all that user code can do is mark
additional nodes as dirty.

Lastly, a note on the new `willReceiveAttrs` hook. The intent of this
hook is for users to be notified that the attributes passed in by their
parent have changed.

When the attributes for a component become dirty, either through an
event or an explicit dirtying operation (i.e. rerender), one of three
things can happen:

1) One of the component’s attributes are passed along to a child
   component. In this case, the child component will become dirtied and
   its `willReceiveAttrs` hook will be invoked.
2) The `willReceiveAttrs` hook of the component mutates its own state
   and `rerender()`s (like React’s `setState()`). In this case, the
   act of re-rendering causes the child component’s `willReceiveAttrs`
   hook to be invoked.
3) None of the above. In this case, the child does not need to have its
   `willReceiveAttrs` hook invoked, because we know that the attr is
   not used in a descendent template, and the user has not updated the
   state used in the template.