Compatibility with legacy Handlebars helpers
This commit makes use of the new `invokeHelper` hook in Handlebars to
invoke helpers created with `Handlebars.registerHelper` with the correct
parameters that they were expecting.

The TLDR is that legacy (original generation) Handlebars helpers expect
to receive their parameters as strings + types, so we track the original
paths through streams and pass those to the helper.