Remove global stream memoization to fix more tests
Memoizing into a local variable doesn't work well the tests since
the tests mutate the value of the global. In theory we could
define a test helper to work around this, but it doesn't seem
like its worth the effort since this path is deprecated and
has a simple upgrade path.